% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testequavar.R
\docType{package}
\name{testequavar}
\alias{testequavar}
\alias{testequavar-package}
\title{testequavar Package}
\description{
Tests the hypothesis  that 2, 3, or 4 population variances are homogeneous or not using bootstrap.
}
\details{
Reference:

Cahoy (2010) <doi:10.1016/j.csda.2010.04.012>
}
\author{
Dexter Cahoy \email{cahoyd@uhd.edu}
}
